refactor: extract JiraExportProvider+Discovery (#638 slice D)
Fourth slice of #638. Biggest single-slice yield — moves 8
discovery methods (validate + fetchProjects + fetchIssueTypes +
5 private request/payload helpers) + 6 wire structs.

Widenings: 6 wire structs (file-scope private → internal). All
methods stay at existing visibility (already public/internal).
Actor stored-prop deps: only 'session' (already widened by
#880/#892).

Line-count delta:
- JiraExportProvider.swift: 850 → 463 (-387 lines)
- JiraExportProvider+Discovery.swift (new): 404 lines
- xcodeproj: +4

Test coverage: JiraExportProviderTests passes unchanged.

Closes #894. Continues #638 after PRs #877/#891/#893.
